+++ dDatabase/GameDatabase/MySQL/Tables/Ugc.cpp
@@ -46,7 +46,7 @@ void MySQLDatabase::RemoveUnreferencedUgcModels() {
 
 void MySQLDatabase::InsertNewUgcModel(
 	std:: stringstream& sd0Data, // cant be const sad
-	const uint32_t blueprintId,
+	const uint64_t blueprintId,
 	const uint32_t accountId,
 	const LWOOBJID characterId) {
 	const std::istream stream(sd0Data.rdbuf());
